feat (ResultView) : 로또 티켓 개수 출력 기능 구현

+++ src/main/java/lottery/domain/LotteryMachine.java
@@ -15,7 +15,7 @@ public class LotteryMachine {
             throw new IllegalArgumentException(ErrorMessages.PURCHASE_PRICE_NOT_ENOUGH);
     }
 
-    public long getLotteryTicketCounts() {
+    public int getLotteryTicketCounts() {
         return lotteryTicketCounts;
     }
 }
